mips: Optimize _andi
In the case where the immediate value is a high mask (bits 31-x set) or
low mask (bits x-0 set), it is faster to clear the unwanted bits by
shifing the bits left/right.

On MIPS32r2 and MIPS64r2, for low masks it is even possible to clear all
bits with one single EXT instruction.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>